Move benchmarks tests into separate folder and cleanup redundancies
