+++ tests/benchmarks/test_compilation.py
@@ -1,56 +1,68 @@
+from contextlib import nullcontext
+
 import numpy as np
 import pytest
 
-import pytensor.tensor as pt
+from pytensor import config, function
 from pytensor.graph.replace import graph_replace
 from pytensor.graph.rewriting import rewrite_graph
 from pytensor.graph.traversal import explicit_graph_inputs
+from pytensor.tensor import (
+    as_tensor,
+    concatenate,
+    exp,
+    grad,
+    log,
+    prod,
+    scalar,
+    sqrt,
+    switch,
+    vector,
+)
+from pytensor.tensor import (
+    sum as pt_sum,
+)
 
 
 def create_radon_model(
     intercept_dist="normal", sigma_dist="halfnormal", centered=False
 ):
     def halfnormal(name, *, sigma=1.0, model_logp):
-        log_value = pt.scalar(f"{name}_log")
-        value = pt.exp(log_value)
+        log_value = scalar(f"{name}_log")
+        value = exp(log_value)
 
-        logp = (
-            -0.5 * ((value / sigma) ** 2) + pt.log(pt.sqrt(2.0 / np.pi)) - pt.log(sigma)
-        )
-        logp = pt.switch(value >= 0, logp, -np.inf)
+        logp = -0.5 * ((value / sigma) ** 2) + log(sqrt(2.0 / np.pi)) - log(sigma)
+        logp = switch(value >= 0, logp, -np.inf)
         model_logp.append(logp + value)
         return value
 
     def normal(name, *, mu=0.0, sigma=1.0, model_logp, observed=None):
-        value = pt.scalar(name) if observed is None else pt.as_tensor(observed)
+        value = scalar(name) if observed is None else as_tensor(observed)
 
         logp = (
-            -0.5 * (((value - mu) / sigma) ** 2)
-            - pt.log(pt.sqrt(2.0 * np.pi))
-            - pt.log(sigma)
+            -0.5 * (((value - mu) / sigma) ** 2) - log(sqrt(2.0 * np.pi)) - log(sigma)
         )
         model_logp.append(logp)
         return value
 
     def lognormal(name, *, mu=0.0, sigma=1.0, model_logp):
         value = normal(name, mu=mu, sigma=sigma, model_logp=model_logp)
-        return pt.exp(value)
+        return exp(value)
 
     def zerosumnormal(name, *, sigma=1.0, size, model_logp):
-        raw_value = pt.vector(f"{name}_zerosum", shape=(size - 1,))
+        raw_value = vector(f"{name}_zerosum", shape=(size - 1,))
         n = raw_value.shape[0] + 1
         sum_vals = raw_value.sum(0, keepdims=True)
-        norm = sum_vals / (pt.sqrt(n) + n)
-        fill_value = norm - sum_vals / pt.sqrt(n)
-        value = pt.concatenate([raw_value, fill_value]) - norm
+        norm = sum_vals / (sqrt(n) + n)
+        fill_value = norm - sum_vals / sqrt(n)
+        value = concatenate([raw_value, fill_value]) - norm
 
         shape = value.shape
-        _full_size = pt.prod(shape)
-        _degrees_of_freedom = pt.prod(shape[-1:].inc(-1))
-        logp = pt.sum(
+        _full_size = prod(shape)
+        _degrees_of_freedom = prod(shape[-1:].inc(-1))
+        logp = pt_sum(
             -0.5 * ((value / sigma) ** 2)
-            - (pt.log(pt.sqrt(2.0 * np.pi)) + pt.log(sigma))
-            * (_degrees_of_freedom / _full_size)
+            - (log(sqrt(2.0 * np.pi)) + log(sigma)) * (_degrees_of_freedom / _full_size)
         )
         model_logp.append(logp)
         return value
@@ -114,15 +126,15 @@ def create_radon_model(
         model_logp=model_logp,
     )
 
-    model_logp = pt.sum([logp.sum() for logp in model_logp])
+    model_logp = pt_sum([logp.sum() for logp in model_logp])
     model_logp = rewrite_graph(
         model_logp, include=("canonicalize", "stabilize"), clone=False
     )
     params = list(explicit_graph_inputs(model_logp))
-    model_dlogp = pt.concatenate([term.ravel() for term in pt.grad(model_logp, params)])
+    model_dlogp = concatenate([term.ravel() for term in grad(model_logp, params)])
 
     size = sum(int(np.prod(p.type.shape)) for p in params)
-    joined_inputs = pt.vector("joined_inputs", shape=(size,))
+    joined_inputs = vector("joined_inputs", shape=(size,))
     idx = 0
     replacement = {}
     for param in params:
@@ -154,3 +166,87 @@ def radon_model_variants():
         for intercept_dist in ("normal", "lognormal")
         for sigma_dist in ("halfnormal", "lognormal")
     ]
+
+
+@pytest.mark.parametrize(
+    "mode, cache",
+    [("C", None), ("CVM", None), ("NUMBA", False), ("NUMBA", True)],
+)
+def test_radon_model_compile_repeatedly_benchmark(mode, cache, radon_model, benchmark):
+    joined_inputs, [model_logp, model_dlogp] = radon_model
+    rng = np.random.default_rng(1)
+    x = rng.normal(size=joined_inputs.type.shape).astype(config.floatX)
+
+    def compile_and_call_once():
+        fn = function(
+            [joined_inputs], [model_logp, model_dlogp], mode=mode, trust_input=True
+        )
+        fn(x)
+
+    ctx = (
+        config.change_flags(numba__cache=cache) if cache is not None else nullcontext()
+    )
+    with ctx:
+        benchmark.pedantic(compile_and_call_once, rounds=5, iterations=1)
+
+
+@pytest.mark.parametrize(
+    "mode, cache",
+    [("C", None), ("CVM", None), ("NUMBA", False), ("NUMBA", True)],
+)
+def test_radon_model_compile_variants_benchmark(
+    mode, cache, radon_model, radon_model_variants, benchmark
+):
+    joined_inputs, [model_logp, model_dlogp] = radon_model
+    rng = np.random.default_rng(1)
+    x = rng.normal(size=joined_inputs.type.shape).astype(config.floatX)
+
+    # Compile base function once to populate the cache
+    fn = function(
+        [joined_inputs], [model_logp, model_dlogp], mode=mode, trust_input=True
+    )
+    fn(x)
+
+    def compile_and_call_once():
+        for joined_inputs, [model_logp, model_dlogp] in radon_model_variants:
+            fn = function(
+                [joined_inputs], [model_logp, model_dlogp], mode=mode, trust_input=True
+            )
+            fn(x)
+
+    ctx = (
+        config.change_flags(numba__cache=cache) if cache is not None else nullcontext()
+    )
+    with ctx:
+        benchmark.pedantic(compile_and_call_once, rounds=1, iterations=1)
+
+
+@pytest.mark.parametrize(
+    "mode, cache",
+    [
+        ("C", None),
+        ("CVM", None),
+        ("CVM_NOGC", None),
+        ("NUMBA", False),
+        ("NUMBA", True),
+    ],
+)
+def test_radon_model_call_benchmark(mode, cache, radon_model, benchmark):
+    joined_inputs, [model_logp, model_dlogp] = radon_model
+
+    real_mode = "CVM" if mode == "CVM_NOGC" else mode
+    ctx = (
+        config.change_flags(numba__cache=cache) if cache is not None else nullcontext()
+    )
+    with ctx:
+        fn = function(
+            [joined_inputs], [model_logp, model_dlogp], mode=real_mode, trust_input=True
+        )
+    if mode == "CVM_NOGC":
+        fn.vm.allow_gc = False
+
+    rng = np.random.default_rng(1)
+    x = rng.normal(size=joined_inputs.type.shape).astype(config.floatX)
+    fn(x)  # warmup
+
+    benchmark(fn, x)
