+++ tests/tensor/test_blas_c.py
@@ -443,94 +443,3 @@ class TestSdotNoFlags(TestCGemvNoFlags):
 
 class TestBlasStridesC(TestBlasStrides):
     mode = mode_blas_opt
-
-
-def test_gemv_vector_dot_perf(benchmark):
-    n = 400_000
-    a = pt.vector("A", shape=(n,))
-    b = pt.vector("x", shape=(n,))
-
-    out = CGemv(inplace=True)(
-        pt.empty((1,)),
-        1.0,
-        a[None],
-        b,
-        0.0,
-    )
-    fn = pytensor.function([a, b], out, accept_inplace=True, trust_input=True)
-
-    rng = np.random.default_rng(430)
-    test_a = rng.normal(size=n)
-    test_b = rng.normal(size=n)
-
-    np.testing.assert_allclose(
-        fn(test_a, test_b),
-        np.dot(test_a, test_b),
-    )
-
-    benchmark(fn, test_a, test_b)
-
-
-@pytest.mark.parametrize(
-    "neg_stride1", (True, False), ids=["neg_stride1", "pos_stride1"]
-)
-@pytest.mark.parametrize(
-    "neg_stride0", (True, False), ids=["neg_stride0", "pos_stride0"]
-)
-@pytest.mark.parametrize("F_layout", (True, False), ids=["F_layout", "C_layout"])
-def test_gemv_negative_strides_perf(neg_stride0, neg_stride1, F_layout, benchmark):
-    A = pt.matrix("A", shape=(512, 512))
-    x = pt.vector("x", shape=(A.type.shape[-1],))
-    y = pt.vector("y", shape=(A.type.shape[0],))
-
-    out = CGemv(inplace=False)(
-        y,
-        1.0,
-        A,
-        x,
-        1.0,
-    )
-    fn = pytensor.function([A, x, y], out, trust_input=True)
-
-    rng = np.random.default_rng(430)
-    test_A = rng.normal(size=A.type.shape)
-    test_x = rng.normal(size=x.type.shape)
-    test_y = rng.normal(size=y.type.shape)
-
-    if F_layout:
-        test_A = test_A.T
-    if neg_stride0:
-        test_A = test_A[::-1]
-    if neg_stride1:
-        test_A = test_A[:, ::-1]
-    assert (test_A.strides[0] < 0) == neg_stride0
-    assert (test_A.strides[1] < 0) == neg_stride1
-
-    # Check result is correct by using a copy of A with positive strides
-    res = fn(test_A, test_x, test_y)
-    np.testing.assert_allclose(res, fn(test_A.copy(), test_x, test_y))
-
-    benchmark(fn, test_A, test_x, test_y)
-
-
-@pytest.mark.parametrize("inplace", (True, False), ids=["inplace", "no_inplace"])
-@pytest.mark.parametrize("n", [2**7, 2**9, 2**13])
-def test_ger_benchmark(n, inplace, benchmark):
-    alpha = pt.dscalar("alpha")
-    x = pt.dvector("x")
-    y = pt.dvector("y")
-    A = pt.dmatrix("A")
-
-    out = alpha * pt.outer(x, y) + A
-
-    fn = pytensor.function(
-        [alpha, x, y, pytensor.In(A, mutable=inplace)], out, trust_input=True
-    )
-
-    rng = np.random.default_rng([2274, n])
-    alpha_test = rng.normal(size=())
-    x_test = rng.normal(size=(n,))
-    y_test = rng.normal(size=(n,))
-    A_test = rng.normal(size=(n, n))
-
-    benchmark(fn, alpha_test, x_test, y_test, A_test)
